// src/userOnboardings/domain/models/userOnboarding.model.ts
export default class UserOnboardingModel {
  id?: number | null = null;
  user_id?: number | null = null;
  onboarding_id?: number | null = null;
  state: number | null = null;
  name?: string | null = null;
  start_date?: string | null = null;
  end_date?: string | null = null;
  color?: string | null = null;
  onboarding_type?: string | null = null;
  created_at?: string | null = null;
  updated_at?: string | null = null;
}